% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.multqrsvm.R
\name{fitted.multqrsvm}
\alias{fitted.multqrsvm}
\title{Fitted values of class "multqrsvm"}
\usage{
\method{fitted}{multqrsvm}(object, ...)
}
\arguments{
\item{object}{An object of class "multqrsvm"}

\item{...}{other arguments.}
}
\value{
A numeric vector of predicted values
}
\description{
Fitted values of class "multqrsvm"
}
