% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.multqrsvm.R
\name{predict.multqrsvm}
\alias{predict.multqrsvm}
\title{Predict an Object of class "multqrsvm"}
\usage{
predict.multqrsvm(model, newdata)
}
\arguments{
\item{model}{An object of class "multqrsvm"}

\item{newdata}{The predictors of the predictable data in an n X m Matrix}
}
\value{
A list of predicted values
}
\description{
Predict an Object of class "multqrsvm"
}
