% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.qrsvm.R
\name{predict.qrsvm}
\alias{predict.qrsvm}
\title{Predict an Object of class "qrsvm"}
\usage{
\method{predict}{qrsvm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "qrsvm"}

\item{newdata}{The predictors of the predictable data in an n X m Matrix}

\item{...}{other arguments.}
}
\value{
A numeric vector of predicted values
}
\description{
Predict an Object of class "qrsvm"
}
