% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.qrsvm.R
\name{fitted.qrsvm}
\alias{fitted.qrsvm}
\title{Fitted values of class "qrsvm"}
\usage{
\method{fitted}{qrsvm}(object, ...)
}
\arguments{
\item{object}{An object of class "qrsvm"}

\item{...}{other arguments.}
}
\value{
A numeric vector of predicted values
}
\description{
Fitted values of class "qrsvm"
}
