% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelMat.R
\name{kernelMat}
\alias{kernelMat}
\title{Modified KernelMatrix function to produce less errors}
\usage{
kernelMat(kernel, x, y = NULL)
}
\arguments{
\item{kernel}{The kernel use from kernlab. The available kernels are
"rbfdot", "polydot", "vanilladot", "tanhdot", "laplacedot", "besseldot" or
"anovadot".}

\item{x}{The data.}

\item{y}{Currently not used.}
}
\description{
Modified KernelMatrix function to produce less errors
}
