% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.multqrsvm.R
\name{predict.multqrsvm}
\alias{predict.multqrsvm}
\title{Predict an Object of class "multqrsvm"}
\usage{
\method{predict}{multqrsvm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \link{multqrsvm}.}

\item{newdata}{The predictors of the predictable data in an n X m Matrix.}

\item{...}{other arguments.}
}
\value{
A list of predicted values.
}
\description{
Predict an Object of class "multqrsvm"
}
