% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qrsvm.R
\name{predict.qrsvm}
\alias{predict.qrsvm}
\title{Predict ann Object oc class "qrsvm"}
\usage{
predict.qrsvm(model, newdata)
}
\arguments{
\item{model}{An object of class "qrsvm"}

\item{newdata}{The predictors of the predictable data in an n X m Matrix}
}
\value{
A numeric vector of predicted values
}
\description{
Predict ann Object oc class "qrsvm"
}
