% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrsvm.R
\name{qrsvm}
\alias{qrsvm}
\title{Fits a quantile regression SVM based on the Pinball Loss}
\usage{
qrsvm(x, y, kernel = "rbfdot", cost = 1, tau = 0.95, sigma = 5,
  degree = 2, scale = 1, offset = 1, order = 1)
}
\arguments{
\item{x}{An n X m matrix containing the predictors (n= number of observatiosn, m = number of predictors).}

\item{y}{The Response onto which the qrsvm shall be fitted.}

\item{kernel}{a string giving the type of kernels from package \link{\code{kernlab}} to use f.e.
"rbfdot" for Radial Basis Function Kernel. All Kernels except "stringdot" supported.}

\item{cost}{The Cost parameter see f.e. package \link{\code{e1071}} and \link{\code{kernlab}}}

\item{tau}{The Quantile that shall be estimated. 0<=tau<=1.}

\item{sigma, degree, scale, offset, order}{A possible tuning parameter for specific Kernelfunctions, see \link{\code{kernlab}}.}
}
\value{
An object of class "qrsvm".
}
\description{
Fits a quantile regression SVM based on the Pinball Loss
}
\details{
There is no preimplemented scaling of the input variables which should be considered beforehand.
  Also optimization is based on "quadprog:solve.QP" function which can be considerably slow compared to
  other SVM implementations.
}
\examples{
# data generation
n <- 200
x <- as.matrix(seq(-1.5, 1.5, length.out = n))
y <- rnorm(n) * (0.3 + abs(sin(x)))

# fit models
mod005 <- qrsvm(x, y, tau = 0.05)
mod050 <- qrsvm(x, y, tau = 0.5)
mod095 <- qrsvm(x, y, tau = 0.95)

# methods
print(mod050)
summary(mod050)
fittedDf <- data.frame(x, fitted05 = fitted(mod005), fitted50 = fitted(mod050),
                       fitted95 = fitted(mod095))
predict(mod050, c(-1, 0, 1))

# graph
library(ggplot2)
ggplot(data.frame(x, y), aes(x, y)) + geom_point() +
  geom_line(aes(x, fitted05, colour = "P05"), fittedDf) +
  geom_line(aes(x, fitted50, colour = "P50"), fittedDf) +
  geom_line(aes(x, fitted95, colour = "P95"), fittedDf) +
  labs(colour = expression(tau))
}
\references{
"Nonparametric Quantile Regression" by I.Takeuchi, Q.V. Le, T. Sears, A.J. Smola (2004)
}
